clippy

+++ core/tauri-runtime-wry/src/lib.rs
@@ -3285,8 +3285,8 @@ fn create_window<T: UserEvent, F: Fn(RawWindow) + Send + 'static>(
         .to_logical::<i32>(scale_factor);
       let screen_size = monitor.size().to_logical::<i32>(scale_factor);
       let monitor_pos = monitor.position().to_logical::<i32>(scale_factor);
-      let x = (screen_size.width as i32 - window_size.width) / 2 + monitor_pos.x;
-      let y = (screen_size.height as i32 - window_size.height) / 2 + monitor_pos.y;
+      let x = (screen_size.width - window_size.width) / 2 + monitor_pos.x;
+      let y = (screen_size.height - window_size.height) / 2 + monitor_pos.y;
 
       window_builder = window_builder.position(x as f64, y as f64);
     }
